+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### API changes
 
+- Provide `DELETE /api/users/me` endpoint
 - Provide `POST /api/users/password_resets` endpoint
 - Provide `POST /api/users/passwords` endpoint
 - Provide `POST /api/users/activation_emails` endpoint